+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed platform providers for `piral-ng` with Angular 20 in non-standalone form (#764)
 - Fixed debug mode deserialization in case of modifications from other security contexts (#775)
+- Removed `container` property from the `unhandled-error` event arguments (#777)
 - Improved dialog order in `piral-modals` (#773) by @pranav-hsg
 - Improved output when installation of packages fails (#759)
 - Improved error message when template resolution fails (#763)